+++ app/models/location.rb
@@ -1,3 +1,5 @@
+require_relative '../../app/helpers/image_cloud_helper'
+
 class Location < ActiveRecord::Base
 
   has_many :items
@@ -31,6 +33,9 @@ class Location < ActiveRecord::Base
   end
 
   def remove_images
-    images.delete_all
+    if images.count > 0
+      ImageCloudHelper.remove_images_by_id(images)
+      images.delete_all
+    end
   end
 end
