+++ app/models/review.rb
@@ -15,6 +15,9 @@ class Review < ActiveRecord::Base
   end
 
   def remove_images
-    images.delete_all
+    if images.count > 0
+      ImageCloudHelper.remove_images_by_id(images)
+      images.delete_all
+    end
   end
 end
